pyproject.toml
[project]
name = "runbook"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Rob Donnelly", email = "Robert.Donnelly@jpl.nasa.gov" }
]
requires-python = ">=3.12"
dependencies = [
    "libtmux>=0.55.0",
    "readchar>=4.2.2",
]

[project.scripts]
runbook = "runbook.cli:main"

[build-system]
requires = ["uv_build>=0.11.1,<0.12.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.9",
    "textual-dev>=1.8.0",
    "ty>=0.0.28",
]
